Split test.py into separate files
