+++ .pre-commit-config.yaml
@@ -38,7 +38,7 @@ repos:
       exclude: |
           (?x)^(
             # Enable these later, avoid bloating this PR
-            bindings/python/test.py|
+            bindings/python/tests/.*.py|
             docs/gen_reference_doc.py|
             examples/run_benchmarks.py|
             fuzzers/tools/generate_initial_corpus.py|
@@ -70,7 +70,7 @@ repos:
       # Avoiding PR bloat
       exclude: |
           (?x)^(
-               bindings/python/test.py|
+               bindings/python/tests/.*.py|
                docs/filter-rst.py|
                docs/gen_reference_doc.py|
                docs/gen_settings_doc.py|
@@ -109,7 +109,7 @@ repos:
       # Avoiding PR bloat
       exclude: |
           (?x)^(
-              bindings/python/test.py|
+              bindings/python/tests/.*.py|
               docs/filter-rst.py|
               docs/gen_reference_doc.py|
               docs/gen_settings_doc.py|
@@ -146,7 +146,7 @@ repos:
     exclude: |
         (?x)^(
             # Enable these later, avoid bloating this PR
-            bindings/python/test.py|
+            bindings/python/tests/.*.py|
             docs/gen_todo.py|
             docs/gen_reference_doc.py|
             docs/gen_settings_doc.py|
