+++ tox.ini
@@ -23,4 +23,4 @@ deps =
 commands =
     {envpython} setup.py bdist_wheel
     {envpython} -m pip install --upgrade --force-reinstall --ignore-installed --find-links=dist python-libtorrent
-    {envpython} -X dev -m pytest --log-level=DEBUG --log-cli-level=DEBUG test.py
+    {envpython} -X dev -m pytest --log-level=DEBUG --log-cli-level=DEBUG
